🔖 chore(pom.xml): remove name tag and update description The name tag has been removed from the pom.xml file as it is not necessary. The description tag has been updated to provide a more accurate description of the project.
🐛 fix(OpenApiConfig.java): fix indentation and line breaks
✨ feat(OpenApiConfig.java): add Request-ID header to requests and responses
The indentation and line breaks were fixed to improve the code's readability. The Request-ID header was added to requests and responses to identify the call. If the header is not passed, it is self-generated, and the ID is returned in the response.

🔥 refactor(application.properties): remove unused property
The `info.application.name` property is no longer used in the application and has been removed from the `application.properties` file.